LoadState: Casting int incorrectly Server: Casting int incorrectly

+++ trunk/src/LoadState.cpp
@@ -114,7 +114,7 @@ void LoadState::waitForReply(void)
             {
                 printf("A packet of length %u containing 0x%08x was received on channel %u.\n",
                     event.packet->dataLength,
-                    (unsigned int)event.packet->data,
+                    (intptr_t)event.packet->data,
                     event.channelID);
                 Player player;
                 memcpy(&player, event.packet->data, sizeof(Player));
